Fixed all the TypeOfX fields that were not being assigned by mistake, and updated list count type to uint32_t instead of uint8_t

+++ Nernel/Include/CLR/MetaData/MethodDefinition.h
@@ -82,7 +82,7 @@ struct _MethodDefinition
     uint32_t SignatureLength;
     const uint8_t* Signature;
     Parameter* ParameterList;
-    uint8_t ParameterListCount;
+    uint32_t ParameterListCount;
 };
 
 const uint8_t* MethodDefinition_Initialize(CLIFile* pFile, const uint8_t* pTableData);
